+++ tests/unit/train/models/test_nemotron_h.py
@@ -56,7 +56,7 @@ def get_model_pairs():
     )
     prime_config._attn_implementation = "flash_attention_2"
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_model = HFNemotronHForCausalLM._from_config(hf_config)
         prime_model = NemotronHForCausalLM._from_config(prime_config)
 
@@ -85,7 +85,7 @@ def test_nemotron_h_mamba_moe_only():
             layer.forward = lambda hidden_states, **kwargs: hidden_states
 
     torch.manual_seed(42)
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, 256, (1, 32))
         position_ids = torch.arange(0, 32).unsqueeze(0)
 
@@ -115,7 +115,7 @@ def test_nemotron_h_reverse():
     hf_config = HFNemotronHConfig(**_BASE, hybrid_override_pattern="ME*E")
     hf_config._attn_implementation = "flash_attention_2"
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         prime_model = NemotronHForCausalLM._from_config(prime_config)
         hf_model = HFNemotronHForCausalLM._from_config(hf_config)
 
@@ -139,7 +139,7 @@ def test_nemotron_h_reverse():
         if isinstance(layer, NemotronHAttentionLayer):
             layer.forward = lambda hidden_states, **kwargs: hidden_states
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, 256, (1, 32))
         position_ids = torch.arange(0, 32).unsqueeze(0)
 
@@ -156,7 +156,7 @@ def test_nemotron_h():
     """Test full model (Mamba + MoE + Attention) produces close outputs."""
     hf_model, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, 256, (1, 32))
         position_ids = torch.arange(0, 32).unsqueeze(0)
