+++ tests/unit/train/models/test_qwen3_5_moe.py
@@ -32,7 +32,7 @@ def get_model_pairs():
         use_grouped_mm=False,
     )
     config._attn_implementation = "flash_attention_2"
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_model = HFQwen3_5MoeForCausalLM._from_config(config)
         prime_model = PrimeRLQwen3_5MoeForCausalLM._from_config(config)
     with torch.no_grad():
@@ -48,7 +48,7 @@ def get_model_pairs():
 def test_qwen3_5_moe():
     hf_model, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -94,7 +94,7 @@ def test_qwen3_5_moe_router_replay():
     """When routed_experts are provided, the model uses them instead of computing routing."""
     _, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, prime_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
