+++ tests/unit/train/models/test_qwen3_moe.py
@@ -32,7 +32,7 @@ def get_model_pairs():
     # But the grad seems to be off in attn because of precision
     # hf_config._attn_implementation = "flash_attention_2"
     hf_config._attn_implementation = "flash_attention_2"
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_model = HFQwen3MoeForCausalLM._from_config(hf_config)
         prime_model = PrimeRLQwen3MoeForCausalLM._from_config(hf_config)
     with torch.no_grad():
@@ -58,7 +58,7 @@ def test_qwen3_moe_attn_only():
     for layer in prime_model.model.layers:
         layer.mlp = _IdentityMLP()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -86,7 +86,7 @@ def test_qwen3_moe_mlp_only():
     for layer in prime_model.model.layers:
         layer.self_attn.forward = foo
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -106,7 +106,7 @@ def test_qwen3_moe_mlp_only():
 def test_qwen3_moe():
     hf_model, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -127,7 +127,7 @@ def test_qwen3_moe_router_replay():
     """When routed_experts are provided, the model uses them instead of computing routing."""
     _, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, prime_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
